backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Salarium backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "bcrypt==3.2.0",
    "fastapi==0.114.1",
    "openpyxl==3.1.5",
    "pandas==2.2.2",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.9.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "tortoise-orm==0.21.5",
    "uvicorn[standard]==0.30.6",
]
